removed warnings, added documentation
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@6155 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/netimport/NIArcView_Loader.h
@@ -75,7 +75,7 @@ private:
     SUMOReal getSpeed(OGRFeature &f, const std::string &edgeid);
 
     /// parses the number of lanes of the edge currently processed
-    size_t getLaneNo(OGRFeature &f,
+    unsigned int getLaneNo(OGRFeature &f,
                      const std::string &edgeid, SUMOReal speed);
 
     /// parses the priority of the edge currently processed
